requirements.txt
flask==1.0.2
Flask-Bcrypt==0.7.1
flask-jwt-extended==3.18.2
Flask-RESTful==0.3.7
mongoengine==0.17.0
mongoengine_goodjson==1.1.6
python-dotenv==0.10.2